# app/Http/Controllers/Api/VacationCareController.php
<?php

namespace App\Http\Controllers\Api;

use App\Http\Controllers\Controller;
use App\Services\GoogleMapsService;
use Illuminate\Http\Request;
use App\Models\User;

class VacationCareController extends Controller
{
    private $mapsService;

    public function __construct(GoogleMapsService $mapsService)
    {
        $this->mapsService = $mapsService;
    }

    /**
     * @OA\Post(
     *     path="/api/vacation-care/search",
     *     summary="Search for nannies in vacation location",
     *     tags={"Bookings"},
     *     @OA\RequestBody(
     *         @OA\JsonContent(
     *             @OA\Property(property="destination", type="string"),
     *             @OA\Property(property="start_date", type="string", format="date"),
     *             @OA\Property(property="end_date", type="string", format="date"),
     *             @OA\Property(property="children_ages", type="array", @OA\Items(type="integer")),
     *             @OA\Property(property="care_schedule", type="object"),
     *             @OA\Property(property="accommodation_type", type="string"),
     *             @OA\Property(property="special_requirements", type="string")
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Successful operation"
     *     )
     * )
     */
    public function searchVacationNannies(Request $request)
    {
        $request->validate([
            'destination' => 'required|string',
            'start_date' => 'required|date|after:today',
            'end_date' => 'required|date|after:start_date',
            'children_ages' => 'required|array',
            'care_schedule' => 'required|array',
        ]);

        // Placeholder: implement search logic using services
        $localNannies = [];
        $travelingNannies = [];
        $destinationInsights = $this->getDestinationInsights($request->destination);

        return response()->json([
            'local_nannies' => $localNannies,
            'traveling_nannies' => $travelingNannies,
            'destination_insights' => $destinationInsights,
            'estimated_local_rates' => $this->estimateLocalRates($request->destination),
        ]);
    }

    private function searchTravelingNannies($origin, $destination, $criteria)
    {
        return User::query()
            ->role('nanny')
            ->where('is_active', true)
            ->where('willing_to_travel', true)
            ->get();
    }

    private function getDestinationInsights($destination)
    {
        $coords = $this->mapsService->geocodeAddress($destination);

        return [
            'average_hourly_rate' => $this->getAverageRate($coords),
            'available_nannies_count' => $this->getAvailableNanniesCount($coords),
            'popular_certifications' => $this->getPopularCertifications($coords),
            'local_regulations' => $this->getLocalChildcareRegulations($coords),
            'emergency_contacts' => $this->getEmergencyContacts($coords),
        ];
    }

    // The following helper methods are placeholders for demonstration
    private function estimateLocalRates($destination)
    {
        return [];
    }

    private function getAverageRate($coords)
    {
        return 0;
    }

    private function getAvailableNanniesCount($coords)
    {
        return 0;
    }

    private function getPopularCertifications($coords)
    {
        return [];
    }

    private function getLocalChildcareRegulations($coords)
    {
        return '';
    }

    private function getEmergencyContacts($coords)
    {
        return [];
    }
}