# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kynderway API",
        "description": "API documentation for Kynderway",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/vacation-care/search": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Search for nannies in vacation location",
                "operationId": "9b8794b135577c81d25b897fdd823bcc",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "destination": {
                                        "type": "string"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "children_ages": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "care_schedule": {
                                        "type": "object"
                                    },
                                    "accommodation_type": {
                                        "type": "string"
                                    },
                                    "special_requirements": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/kyc/verify-document": {
            "post": {
                "summary": "Upload and verify identity document",
                "operationId": "2c45ad04695349d2dd5ae10880a6bb41",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "document": {
                                        "type": "file"
                                    },
                                    "document_type": {
                                        "type": "string",
                                        "enum": [
                                            "passport",
                                            "drivers_license",
                                            "national_id"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Document uploaded and verification in progress"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT based authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}